# docs/swagger.yaml
basePath: /api/v2
definitions:
...
  github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse:
    properties:
      code:
        type: integer
      details:
        type: object
      message:
        type: string
    type: object
...
  github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ListingAgendaDetailResponse:
    properties:
      entries:
        items:
          $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ScheduleEntryResponse'
        type: array
      pagination:
        $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.PaginationResponse'
    type: object
...
  github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.PaginationResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
...
  github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ScheduleAvailabilityResponse:
    properties:
      pagination:
        $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.PaginationResponse'
      slots:
        items:
          $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ScheduleAvailabilitySlotResponse'
        type: array
      timezone:
        type: string
    type: object
  github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ScheduleAvailabilitySlotResponse:
    properties:
      endsAt:
        type: string
      startsAt:
        type: string
    type: object
  github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ScheduleBlockEntryRequest:
    properties:
      endsAt:
        type: string
      entryType:
        type: string
      listingId:
        type: integer
      reason:
        type: string
      startsAt:
        type: string
      timezone:
        example: America/Sao_Paulo
        type: string
    required:
    - endsAt
    - entryType
    - listingId
    - startsAt
    - timezone
    type: object
  github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ScheduleBlockEntryResponse:
    properties:
      entry:
        $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ScheduleEntryResponse'
    type: object
  github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ScheduleBlockEntryUpdateRequest:
    properties:
      endsAt:
        type: string
      entryId:
        type: integer
      entryType:
        type: string
      listingId:
        type: integer
      reason:
        type: string
      startsAt:
        type: string
      timezone:
        example: America/Sao_Paulo
        type: string
    required:
    - endsAt
    - entryId
    - entryType
    - listingId
    - startsAt
    - timezone
    type: object
  github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ScheduleDeleteEntryRequest:
    properties:
      entryId:
        type: integer
      listingId:
        type: integer
    required:
    - entryId
    - listingId
    type: object
  github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ScheduleEntryResponse:
    properties:
      blocking:
        type: boolean
      endsAt:
        type: string
      entryType:
        type: string
      id:
        type: integer
      photoBookingId:
        type: integer
      reason:
        type: string
      startsAt:
        type: string
      timezone:
        type: string
      visitId:
        type: integer
    type: object
...
host: api.gca.dev.br
info:
  contact:
    email: projeto.toq@gmail.com
    name: TOQ Development Team
    url: https://github.com/projeto-toq/toq_server
  description: TOQ Server - Real Estate HTTP API Server with hexagonal architecture.
  license:
    name: MIT
    url: https://github.com/projeto-toq/toq_server/blob/main/LICENSE
  termsOfService: http://swagger.io/terms/
  title: TOQ Server API
  version: "2.0"
paths:
...
  /schedules/listing/availability:
    get:
      description: Returns the available slots for a listing within the provided time
        range.
      parameters:
      - description: Listing identifier
        in: query
        name: listingId
        required: true
        type: integer
      - description: Start of time range (RFC3339)
        in: query
        name: rangeFrom
        type: string
      - description: End of time range (RFC3339)
        in: query
        name: rangeTo
        type: string
      - description: Desired slot duration in minutes
        in: query
        name: slotDurationMinute
        type: integer
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - default: America/Sao_Paulo
        description: Timezone identifier (IANA)
        in: query
        name: timezone
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ScheduleAvailabilityResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List listing availability
      tags:
      - Schedules
  /schedules/listing/block:
    delete:
      consumes:
      - application/json
      description: Removes a blocking or temporary block entry from a listing agenda.
      parameters:
      - description: Block entry deletion payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ScheduleDeleteEntryRequest'
        x-example: '{"entryId":5021,"listingId":3241}'
      produces:
      - application/json
      responses:
        "204":
          description: Entry deleted successfully
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a blocking entry
      tags:
      - Schedules
    post:
      consumes:
      - application/json
      description: Creates a blocking or temporary blocking time range for a listing
        agenda.
      parameters:
      - description: Block entry payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ScheduleBlockEntryRequest'
        x-example: '{"listingId":3241,"entryType":"BLOCK","startsAt":"2025-06-15T09:00:00-03:00","endsAt":"2025-06-15T11:00:00-03:00","reason":"Janela
          de manutencao","timezone":"America/Sao_Paulo"}'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ScheduleBlockEntryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a blocking entry
      tags:
      - Schedules
    put:
      consumes:
      - application/json
      description: Updates an existing blocking or temporary block entry for a listing
        agenda.
      parameters:
      - description: Block entry update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ScheduleBlockEntryUpdateRequest'
        x-example: '{"entryId":5021,"listingId":3241,"entryType":"TEMP_BLOCK","startsAt":"2025-06-20T09:00:00-03:00","endsAt":"2025-06-20T12:00:00-03:00","reason":"Janela
          de manutencao","timezone":"America/Sao_Paulo"}'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ScheduleBlockEntryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a blocking entry
      tags:
      - Schedules
  /schedules/listing/detail:
    get:
      description: Returns all agenda entries for a specific listing owned by the
        authenticated user.
      parameters:
      - description: Listing identifier
        in: query
        name: listingId
        required: true
        type: integer
      - description: Start of time range (RFC3339)
        in: query
        name: rangeFrom
        type: string
      - description: End of time range (RFC3339)
        in: query
        name: rangeTo
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - default: America/Sao_Paulo
        description: Timezone identifier (IANA)
        in: query
        name: timezone
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ListingAgendaDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List agenda entries for a listing
      tags:
      - Schedules
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"